requirements.txt
anthropic>=0.50.0
google-auth>=2.37.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.160.0
